operator/clustermesh: add endpoint slice synchronization
This commit adds a new clustermesh module inside the operator which pull data
from Cilium's clustermesh components and unlike the Cilium agent will
act on the whole cluster level instead of the node level.

This commit primarily adds optional endpoint slice synchronization from
remote clusters. Having endpoint slices of others clusters is
useful for non Cilium controllers that would need this information
(for instance CoreDNS, non Cilium ingress controller). This
is also a base to support the endpoint slice synchronization from
Multi-Cluster Services API (KEP-1645) and it is using a mcs-api label to
identify the source cluster.

The endpoint slice synchronization is heavily based on existing
Kubernetes code from the endpoint slice controller/reconciler but adapted
to the Cilium use case. For instance, it has the same logic of
splitting endpoints in multiple EndpointSlices, reusing them etc. The
Topology features were removed as this is not something supported in
Cilium clustermesh though.

The endpoint slice synchronization is guarded by a new annotation
`cilium.io/global-sync-endpoint-slices` as enabling it on global services
with high churn may be a scalability concern on all the etcd and/or kube-apiserver
of all the Kubernetes cluster which involve the same global services with the
endpoint slice sync enabled. Note that, again, the same logic as the official
Kubernetes endpoint slice controller have been included here as well to
alleviate those scalability issues.

Related to GitHub issue #27902

Signed-off-by: Arthur Outhenin-Chalandre <arthur@cri.epita.fr>