gateway-api: add ServiceImport support
Add ServiceImport supports in Cilium Gateway API. The only implementation
supported by this commit is the one where ServiceImport reference
an existing "derived" Service with the annotation
`multicluster.kubernetes.io/derived-service`. This way of implementing
`ServiceImport` make it as a "dummy" object while the real logic is still
with the Service object. This implementation of MCS API is not enforced
by the MCS API KEP but other implementations are very unlikely as it needs
to modify kube-proxy (and/or other kube-proxy replacements) to support
`ServiceImport` objects natively. It's also the recommended approach in
the mcs-api repo "reference implementation" and what's planned for
Cilium cluster mesh (see issue #27902).

Since we do not support any ServiceImport that doesn't have a derived
annotation the support in Cilium Gateway API is mainly about making sure
the annotation and the derived service actually exist and swapping
ServiceImport by the derived Service right before ingesting it in Envoy.

Signed-off-by: Arthur Outhenin-Chalandre <arthur@cri.epita.fr>